Extract AccessEntity parsing from DiskAccessStorage
